libxbmc: create an entrypoint so that xbmc and other apps can run the same way
