+++ xbmc/xbmc.cpp
@@ -19,77 +19,15 @@
  *
  */
 
-
-// XBMC
-//
-// libraries:
-//   - CDRipX   : doesnt support section loading yet
-//   - xbfilezilla : doesnt support section loading yet
-//
-
-#include "system.h"
-#include "settings/AppParamParser.h"
-#include "settings/AdvancedSettings.h"
-#include "FileItem.h"
 #include "Application.h"
-#include "PlayListPlayer.h"
-#include "utils/log.h"
-#ifdef _LINUX
-#include <sys/resource.h>
-#include <signal.h>
-#endif
-#if defined(TARGET_DARWIN_OSX)
-  #include "Util.h"
-  // SDL redefines main as SDL_main 
-  #ifdef HAS_SDL
-    #include <SDL/SDL.h>
-  #endif
-#endif
-#ifdef HAS_LIRC
-#include "input/linux/LIRC.h"
-#endif
-#include "XbmcContext.h"
-
-int main(int argc, char* argv[])
+#include "settings/AdvancedSettings.h"
+extern "C" int XBMC_Run(bool renderGUI)
 {
-  // set up some xbmc specific relationships
-  XBMC::Context context;
-
   int status = -1;
-  bool renderGUI = true;
-  //this can't be set from CAdvancedSettings::Initialize() because it will overwrite
-  //the loglevel set with the --debug flag
-#ifdef _DEBUG
-  g_advancedSettings.m_logLevel     = LOG_LEVEL_DEBUG;
-  g_advancedSettings.m_logLevelHint = LOG_LEVEL_DEBUG;
-#else
-  g_advancedSettings.m_logLevel     = LOG_LEVEL_NORMAL;
-  g_advancedSettings.m_logLevelHint = LOG_LEVEL_NORMAL;
-#endif
-  CLog::SetLogLevel(g_advancedSettings.m_logLevel);
 
-#ifdef _LINUX
-#if defined(DEBUG)
-  struct rlimit rlim;
-  rlim.rlim_cur = rlim.rlim_max = RLIM_INFINITY;
-  if (setrlimit(RLIMIT_CORE, &rlim) == -1)
-    CLog::Log(LOGDEBUG, "Failed to set core size limit (%s)", strerror(errno));
-#endif
-  // Prevent child processes from becoming zombies on exit if not waited upon. See also Util::Command
-  struct sigaction sa;
-  memset(&sa, 0, sizeof(sa));
+  if (!g_advancedSettings.Initialized())
+    g_advancedSettings.Initialize();
 
-  sa.sa_flags = SA_NOCLDWAIT;
-  sa.sa_handler = SIG_IGN;
-  sigaction(SIGCHLD, &sa, NULL);
-#endif
-  setlocale(LC_NUMERIC, "C");
-  g_advancedSettings.Initialize();
-  
-#ifndef _WIN32
-  CAppParamParser appParamParser;
-  appParamParser.Parse((const char **)argv, argc);
-#endif
   if (!g_application.Create())
   {
     fprintf(stderr, "ERROR: Unable to create application. Exiting\n");
